feat(FormlyFormController): Exposes FormlyFormController globally
For the formly plugin I am writing, I want to reuse the logic in the FormlyFormController with a
different directive. This change makes the controller available to third party code via the
$controller service.
