+++ src/directives/formly-form.test.js
@@ -4,9 +4,8 @@
 /* eslint max-nested-callbacks:0 */
 import testUtils from '../test.utils.js'
 import angular from 'angular-fix'
-import _ from 'lodash'
 
-const {getNewField, input, basicForm, shouldWarnWithLog} = testUtils
+const {getNewField, input, basicForm} = testUtils
 
 describe('formly-form', () => {
   let $compile, formlyConfig, scope, el, $timeout
@@ -901,107 +900,6 @@ describe('formly-form', () => {
         const autoCompleteFixEl = el[0].querySelector('[autocomplete="address-level4"]')
         expect(autoCompleteFixEl).to.exist
       }))
-    })
-
-    describe(`fieldTransform`, () => {
-      beforeEach(() => {
-        formlyConfig.extras.fieldTransform = fieldTransform
-      })
-
-      it(`should give a deprecation warning when formlyConfig.extras.fieldTransform is a function rather than an array`, inject(($log) => {
-        shouldWarnWithLog(
-          $log,
-          [
-            'Formly Warning:',
-            'fieldTransform as a function has been deprecated.',
-            /Attempted for formlyConfig.extras/,
-          ],
-          compileAndDigest
-        )
-      }))
-
-      it(`should give a deprecation warning when options.fieldTransform function rather than an array`, inject(($log) => {
-        formlyConfig.extras.fieldTransform = undefined
-        scope.fields = [getNewField()]
-        scope.options.fieldTransform = fields => fields
-        shouldWarnWithLog(
-          $log,
-          [
-            'Formly Warning:',
-            'fieldTransform as a function has been deprecated.',
-            'Attempted for form',
-          ],
-          compileAndDigest
-        )
-      }))
-
-      it(`should throw an error if something is passed in and nothing is returned`, () => {
-        scope.fields = [getNewField()]
-        scope.options.fieldTransform = function() {
-          // I return nothing...
-        }
-        expect(() => compileAndDigest()).to.throw(/^Formly Error: fieldTransform must return an array of fields/)
-      })
-
-      it(`should allow you to transform field configuration`, () => {
-        scope.options.fieldTransform = fieldTransform
-        const spy = sinon.spy(scope.options, 'fieldTransform')
-        doExpectations(spy)
-      })
-
-      it(`should use formlyConfig.extras.fieldTransform when not specified on options`, () => {
-        const spy = sinon.spy(formlyConfig.extras, 'fieldTransform')
-        doExpectations(spy)
-      })
-
-      it(`should allow you to use an array of transform functions`, () => {
-        scope.fields = [getNewField({
-          customThing: 'foo',
-          otherCustomThing: {
-            whatever: '|-o-|',
-          }})]
-        scope.options.fieldTransform = [fieldTransform]
-        expect(() => compileAndDigest()).to.not.throw()
-
-        const field = scope.fields[0]
-        expect(field).to.have.deep.property('data.customThing')
-        expect(field).to.have.deep.property('data.otherCustomThing')
-      })
-
-      function doExpectations(spy) {
-        const originalFields = [{
-          key: 'keyProp',
-          template: '<hr />',
-          customThing: 'foo',
-          otherCustomThing: {
-            whatever: '|-o-|',
-          },
-        }]
-        scope.fields = originalFields
-        compileAndDigest()
-        expect(spy).to.have.been.calledWith(originalFields, scope.model, scope.options, scope.form)
-        const field = scope.fields[0]
-
-        expect(field).to.not.have.property('customThing')
-        expect(field).to.not.have.property('otherCustomThing')
-        expect(field).to.have.deep.property('data.customThing')
-        expect(field).to.have.deep.property('data.otherCustomThing')
-      }
-
-      function fieldTransform(fields) {
-        const extraKeys = ['customThing', 'otherCustomThing']
-        return _.map(fields, field => {
-          const newField = {data: {}}
-          _.each(field, (val, name) => {
-            if (_.includes(extraKeys, name)) {
-              newField.data[name] = val
-            } else {
-              newField[name] = val
-            }
-          })
-          return newField
-        })
-      }
     })
 
     describe(`data`, () => {
