+++ src/index.test.js
@@ -9,6 +9,7 @@ import './directives/formly-custom-validation.test'
 import './directives/formly-field.test'
 import './directives/formly-focus.test'
 import './directives/formly-form.test'
+import './directives/formly-form.controller.test'
 import './run/formlyCustomTags.test'
 import './run/formlyNgModelAttrsManipulator.test'
 import './other/utils.test'
